requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
Babel==2.9.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.8.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro==1.5.0
distro-info==1.0
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.45.0
django-filter==21.1
django-isbn-field==0.5.2
django-money==2.0.1
django-on-heroku==1.1.2
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
duplicity==0.8.17
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
httplib2==0.18.1
idna==3.2
importlib-metadata==1.6.0
inflection==0.5.1
itypes==1.2.0
jeepney==0.6.0
Jinja2==3.0.1
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.16.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.9
oauthlib==3.1.1
olefile==0.46
openapi-codec==1.3.2
packaging==21.0
paramiko==2.7.2
pbr==5.6.0
pexpect==4.8.0
Pillow==8.1.2
protobuf==3.12.4
psycopg2-binary==2.9.1
py-moneyed==1.2
pycairo==1.16.2
pycparser==2.20
pycups==2.0.1
Pygments==2.10.0
PyGObject==3.38.0
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt===2.2.0-ubuntu0.21.04.1
python-dateutil==2.8.1
python-debian==0.1.39
python-stdnum==1.16
python3-openid==3.2.0
pytz==2021.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.2
systemd-python==234
testresources==2.0.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.6
usb-creator==0.3.7
virtualenv==20.4.0+ds
wadllib==1.3.5
whitenoise==5.3.0
xdg==5
xkit==0.0.0
youtube-dl==2021.4.7
zipp==1.0.0
